# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "description": "A single incident signal to assess.",
        "properties": {
          "message": {
            "description": "Raw event message to assess.",
            "maxLength": 8192,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/SourceType",
            "default": "unknown",
            "description": "Signal origin. One of: application, infrastructure, network, iam, audit, unknown."
          }
        },
        "required": [
          "message"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
      "AnalyzeResponse": {
        "description": "Structured incident context produced by the Sentinel pipeline.",
        "properties": {
          "analysis_scope": {
            "title": "Analysis Scope",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "integer"
          },
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "title": "Evidence",
            "type": "array"
          },
          "explain": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explain"
          },
          "feedback": {
            "additionalProperties": true,
            "title": "Feedback",
            "type": "object"
          },
          "historical_context": {
            "additionalProperties": true,
            "title": "Historical Context",
            "type": "object"
          },
          "hypothesis": {
            "title": "Hypothesis",
            "type": "string"
          },
          "impact": {
            "additionalProperties": true,
            "title": "Impact",
            "type": "object"
          },
          "impact_feedback_endpoint": {
            "title": "Impact Feedback Endpoint",
            "type": "string"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "monitored_object": {
            "title": "Monitored Object",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          },
          "recommended_action": {
            "title": "Recommended Action",
            "type": "string"
          },
          "recommended_steps": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Steps",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "risk_score": {
            "title": "Risk Score",
            "type": "number"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "incident_id",
          "analysis_scope",
          "event_type",
          "monitored_object",
          "message",
          "source",
          "risk_score",
          "severity",
          "priority",
          "confidence",
          "recommended_action",
          "evidence",
          "context",
          "historical_context",
          "impact",
          "hypothesis",
          "recommended_steps",
          "feedback",
          "tier",
          "impact_feedback_endpoint"
        ],
        "title": "AnalyzeResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "SourceType": {
        "description": "Origin of an ingested signal — drives source-aware risk weighting.",
        "enum": [
          "application",
          "infrastructure",
          "network",
          "iam",
          "audit",
          "unknown"
        ],
        "title": "SourceType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "## RAVEN Incident Intelligence API\n\nTwo independent scoring engines are available:\n\n| Endpoint | Engine | Score range | Severity bands |\n|---|---|---|---|\n| `POST /evaluate` | Rule-based (legacy) | Integer **0 – 100** | LOW / MEDIUM / HIGH |\n| `POST /v1/analyze` | Sentinel 4-layer pipeline | Float **0.0 – 1.0** | LOW / MEDIUM / HIGH / CRITICAL |\n\nAll protected endpoints require the `X-API-Key` header.\n\nAll error responses share the envelope `{\"error\": \"...\", \"code\": \"...\"}`.",
    "title": "RAVEN Incident Intelligence API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/evaluate": {
      "post": {
        "description": "Rule-based scoring engine. Returns an **integer** risk score (0 – 100) and a level band.\n\n**Score thresholds:** LOW < 40 · MEDIUM 40–69 · HIGH ≥ 70\n\n**Scoring rules:**\n- `action=login_failed` → +30\n- `attempts ≥ 3` → +40\n- Known suspicious IP → +10\n\nFor pipeline-based multi-factor scoring with float scores (0.0 – 1.0), four severity bands, and recurrence detection, use `POST /v1/analyze`.",
        "operationId": "evaluate_event_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "high_risk": {
                  "summary": "High-risk login attempt (score 80, HIGH)",
                  "value": {
                    "action": "login_failed",
                    "attempts": 5,
                    "ip": "192.168.1.1",
                    "message": "failed login from suspicious IP",
                    "source": "application",
                    "user_id": "user_001"
                  }
                },
                "low_risk": {
                  "summary": "Low-risk page view (score 0, LOW)",
                  "value": {
                    "action": "page_view",
                    "attempts": 0,
                    "ip": "8.8.8.8",
                    "message": "user viewed dashboard",
                    "source": "application",
                    "user_id": "user_002"
                  }
                }
              },
              "schema": {
                "properties": {
                  "action": {
                    "description": "Event action (e.g. login_failed, page_view).",
                    "type": "string"
                  },
                  "attempts": {
                    "default": 0,
                    "description": "Number of attempts.",
                    "type": "integer"
                  },
                  "ip": {
                    "description": "Source IP address.",
                    "type": "string"
                  },
                  "message": {
                    "description": "Event description.",
                    "type": "string"
                  },
                  "source": {
                    "default": "unknown",
                    "description": "Signal origin.",
                    "type": "string"
                  },
                  "user_id": {
                    "description": "User identifier.",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Evaluate Event Evaluate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Score an event (rule-based legacy engine)",
        "tags": [
          "evaluate"
        ]
      }
    },
...
    "/v1/analyze": {
      "post": {
        "description": "Normalizes and scores one incident signal, then returns structured evidence, an investigation hypothesis, context, and suggested next steps. The result supports human investigation and is not an automated verdict.",
        "operationId": "analyze_v1_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "synthetic_normal_operation": {
                  "summary": "Synthetic normal operation",
                  "value": {
                    "message": "Background job completed successfully",
                    "source": "application"
                  }
                },
                "synthetic_service_failure": {
                  "summary": "Synthetic service failure",
                  "value": {
                    "message": "Checkout API returned HTTP 503 after a configuration change in the synthetic staging environment",
                    "source": "application"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Analyze an incident signal",
        "tags": [
          "v1"
        ]
      }
    }
  }
}